Create a file for get_test_info
